Merging following revisions to trunk
[913]
Refactoring Dispatcher and Scaffold classes.
Removed Dispatcher::scaffoldView() and added it to
Scaffold::_scaffoldView() less coupling between the 2 classes now

[914]
Adding settings to fix issue when using get_class() in PHP 4.
If value of var $casedClas is set in controller this name will be Inflected to the proper table name if underscores are used
in the database.

[915]
Added suggestion from ticket #220

git-svn-id: https://svn.cakephp.org/repo/trunk/cake@917 3807eeeb-6ff5-0310-8944-8be069107fe0